4_Retraining_MLFlow/requirement.txt
mlflow==2.10.0
pandas==2.1.2
scikit-learn==1.3.2
xgboost==1.7.6
joblib==1.3.2
PyYAML==6.0.1
requests==2.31.0
boto3==1.34.0